Fixing voice and client name
When voice was not defined it was causing an error. Client name was using the remote address of the http request which was not usable in other parts of the code. I switched it to using the memory pointer of the websocket connection to be able to be used globally in the program to get the client name.